# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Snow School Scheduler API
  description: スキー・スノーボードスクールのシフト管理システム API
  version: 1.0.0
  contact:
    name: Snow School Scheduler
servers:
  - url: http://localhost:3000
    description: 開発環境
  - url: https://snow-school-scheduler-api.example.com
    description: 本番環境
paths:
...
  /api/departments/[id]:
    parameters:
      - $ref: '#/components/parameters/Id'

    get:
      summary: 部門詳細取得
      description: 指定された部門の詳細情報を取得
      tags:
        - 部門管理
      responses:
        '200':
          description: 部門詳細
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Department'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/certifications:
    get:
      summary: 資格一覧取得
      description: 全資格の一覧を取得
      tags:
        - 資格管理
      responses:
        '200':
          description: 資格一覧
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Certification'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      summary: 資格作成
      description: 新しい資格を作成
      tags:
        - 資格管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificationInput'
      responses:
        '201':
          description: 資格作成成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Certification'
                      message:
                        type: string
                        example: 'Certification created successfully'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/auth/invitations/{token}:
    parameters:
      - name: token
        in: path
        required: true
        description: 招待トークン文字列
        schema:
          type: string

    delete:
      summary: 招待トークン無効化
      description: 管理者または作成者が招待トークンを無効化します。
      tags:
        - 招待管理
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 招待トークンの無効化に成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InvitationDeactivation'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: すでに無効化済み
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              example:
                success: false
                error: 'Invitation token is already inactive'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/auth/invitations/active:
    get:
      summary: 最新の有効な招待トークン取得
      description: 現在利用可能な最新の招待トークン情報を返します。
      tags:
        - 招待管理
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 有効な招待トークン情報
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ActiveInvitation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/auth/invitations/{token}/verify:
    parameters:
      - name: token
        in: path
        required: true
        description: 招待トークン文字列
        schema:
          type: string

    get:
      summary: 招待トークン検証
      description: 認証なしで招待トークンの有効性を検証します。
      tags:
        - 招待管理
      responses:
        '200':
          description: 招待トークンが有効
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InvitationValidation'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: トークンは存在したが現在は利用不可
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              example:
                success: false
                error: 'Expired invitation token'
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      summary: シフト更新
      description: 指定されたシフトを更新
      tags:
        - シフト管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftInput'
      responses:
        '200':
          description: シフト更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ShiftWithStats'
                      message:
                        type: string
                        example: 'Shift operation completed successfully'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      summary: シフト削除
      description: 指定されたシフトを削除
      tags:
        - シフト管理
      responses:
        '204':
          description: 削除成功（レスポンス本文なし）
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: リソースのID
      schema:
        type: integer
  responses:
    NotFound:
      description: リソースが見つかりません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error: 'Resource not found'
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error: 'Validation failed'
    ServerError:
      description: サーバーエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error: 'Internal server error'
    Unauthorized:
      description: 認証が必要です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error: 'Authentication token required'
    Forbidden:
      description: 権限が不足しています
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error: 'Insufficient permissions. Admin or Manager role required.'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DepartmentSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: 'スキー'
    InstructorSummary:
      type: object
      properties:
        id:
          type: integer
        lastName:
          type: string
          example: '山田'
        firstName:
          type: string
          example: '太郎'
        status:
          type: string
          enum: [ACTIVE, INACTIVE, RETIRED]
          default: ACTIVE
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
          example: null
    ApiListResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
        count:
          type: integer
          example: 10
        error:
          type: string
          nullable: true
          example: null
    ApiErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Invalid invitation token'

    # 部門関連
    Department:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
          example: 'ski'
        name:
          type: string
          example: 'スキー'
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Certification:
      type: object
      properties:
        id:
          type: integer
        departmentId:
          type: integer
        name:
          type: string
          example: 'スキー指導員'
        shortName:
          type: string
          example: '指導員'
        organization:
          type: string
          example: 'SAJ'
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        department:
          $ref: '#/components/schemas/DepartmentSummary'
    CertificationInput:
      type: object
      required:
        - departmentId
        - name
        - shortName
        - organization
      properties:
        departmentId:
          type: integer
        name:
          type: string
          example: 'スキー指導員'
        shortName:
          type: string
          example: '指導員'
        organization:
          type: string
          example: 'SAJ'
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
          default: true
...
    ShiftType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: 'レッスン'
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        departmentId:
          type: integer
        shiftTypeId:
          type: integer
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        department:
          $ref: '#/components/schemas/Department'
        shiftType:
          $ref: '#/components/schemas/ShiftType'
    ShiftInput:
      type: object
      required:
        - date
        - departmentId
        - shiftTypeId
      properties:
        date:
          type: string
          format: date
        departmentId:
          type: integer
        shiftTypeId:
          type: integer
        description:
          type: string
          nullable: true
        assignedInstructorIds:
          type: array
          items:
            type: integer
          description: '割り当てるインストラクターIDの配列'
          example: [1, 2, 3]
    ShiftWithStats:
      allOf:
        - $ref: '#/components/schemas/Shift'
        - type: object
          properties:
            assignments:
              type: array
              items:
                $ref: '#/components/schemas/ShiftAssignment'
            assignedCount:
              type: integer
              description: 現在割り当てられているインストラクター数

    # シフト割り当て関連
    ShiftAssignment:
      type: object
      properties:
        id:
          type: integer
        shiftId:
          type: integer
        instructorId:
          type: integer
        assignedAt:
          type: string
          format: date-time
        instructor:
          $ref: '#/components/schemas/InstructorSummary'
...
    InvitationDeactivation:
      type: object
      properties:
        message:
          type: string
          example: 'Invitation token deactivated successfully'
        token:
          type: string
        deactivatedAt:
          type: string
          format: date-time
        deactivatedBy:
          type: string
    ActiveInvitation:
      type: object
      properties:
        token:
          type: string
        description:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean
        maxUses:
          type: integer
          nullable: true
        usageCount:
          type: integer
        remainingUses:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
    InvitationValidation:
      type: object
      required:
        - isValid
      properties:
        isValid:
          type: boolean
          example: true
...